Adapt baggers page to the new model
In https://github.com/brownbaglunch/bblfr_data/pull/66 we meant to change the model.

This PR (WIP) takes care of the change.

It will help to make progress in baby steps:

* update the model: https://github.com/brownbaglunch/bblfr_data/pull/66
* update the website to the new model: this PR
* rewrite the website using elasticsearch as the search backend

Note that to test it, you need to generate `../bblfr_data/rewrite/baggers-new.js`.
Then uncomment the `local` part and comment the `remote` part.