Fix locations map
We are using an old URL which is giving back a 404 error.
I'm just fixing the URL to the `jquery.ui.map.js`.

But the fix might not be complete as google maps is now displaying a "For development purposes only" message.
We might have to fix that as well.